Remove mbstring from PHP extensions in CI setup
The mbstring extension has been removed from the list of PHP extensions installed by default in the CI pipeline. This change may be due to a decision to manage dependencies differently or because mbstring is no longer required for the current project setup. Ensure that any functionality depending on mbstring is either refactored or handled by other means.